backend/requirements.txt
# ── Core Framework ────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# ── Auth ──────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.1
httpx==0.27.0

# ── AI / ML ───────────────────────────────────────────────
openai==1.35.7
sentence-transformers==3.0.1
spacy==3.7.5
chromadb==0.5.3
langchain==0.2.6
langchain-openai==0.1.14

# ── Whisper (video assessment) ────────────────────────────
openai-whisper==20231117
ffmpeg-python==0.2.0

# ── Utilities ─────────────────────────────────────────────
python-dotenv==1.0.1
aiofiles==24.1.0
numpy==1.26.4
scikit-learn==1.5.1
# ── PDF Processing ────────────────────────────────────────
PyMuPDF==1.24.5
pdfminer.six==20231228
fpdf2==2.8.1
reportlab==4.0.7
